Added Redux store for fetching userDetails, preventing multiple redundant calls across multiple components.
