+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@reduxjs/toolkit": "^1.5.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
     "@testing-library/user-event": "^7.2.1",
@@ -20,6 +21,7 @@
     "react-ga": "^2.7.0",
     "react-https-redirect": "^1.1.0",
     "react-query": "^2.23.0",
+    "react-redux": "^7.2.2",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.3"
   },
@@ -58,6 +60,7 @@
     "eslint-plugin-react": "^7.21.4",
     "eslint-plugin-react-hooks": "^4.1.2",
     "jest-when": "^2.7.2",
+    "redux-mock-store": "^1.5.4",
     "wait-for-expect": "^3.0.2"
   },
   "jest": {
