+++ src/pages/HomePage/HomePage.js
@@ -1,34 +1,22 @@
 import React from "react";
-import { useQuery } from "react-query";
+import { useSelector } from "react-redux";
 
-import axios from "axios";
 import { Divider } from "antd";
 import { Div } from "lemon-reset";
-import apiRoot from "../../apiRoot";
 import GenericPage from "../GenericPage/GenericPage";
 import FeedbackGroupsPage from "../FeedbackGroupsPage/FeedbackGroupsPage";
 import LoadingSpinner from "../../components/LoadingSpinner/LoadingSpinner";
 import Faq from "../../components/Faq/Faq";
 import LandingPitch from "../../components/LandingPitch/LandingPitch";
+import { selectUserDetailsData, selectUserDetailsIsLoading } from "../../reducers/userDetailsSlice";
 
 type Props = {
     isMobile: boolean,
 };
 
-const USER_DETAILS_QUERY = `query UserDetails {
-  userDetails {
-    username
-    rating
-  }
-}`;
-
 const HomePage = ({ isMobile }: Props) => {
-    const { isLoading, data } = useQuery(
-        [USER_DETAILS_QUERY],
-        () => axios.post(`${apiRoot}/graphql/`, {
-            query: USER_DETAILS_QUERY,
-        }).then((response) => response.data.data.userDetails),
-    );
+    const data = useSelector(selectUserDetailsData);
+    const isLoading = useSelector(selectUserDetailsIsLoading);
 
     if (isLoading) {
         return (
