+++ src/pages/TrackUrlHelpPage/TrackUrlHelpPage.test.js
@@ -2,6 +2,8 @@ import React from "react";
 import { shallow } from "enzyme";
 import TrackUrlHelpPage from "./TrackUrlHelpPage";
 
+jest.mock("react-redux");
+
 describe("TrackUrlHelpPage", () => {
     it("renders a GenericPage containing a TrackUrlHelp", () => {
         const wrapper = shallow(
