+++ src/pages/UserSettingsPage/UserSettingsPage.test.js
@@ -1,10 +1,14 @@
 import React from "react";
 import { mount } from "enzyme";
 import axios from "axios";
-import waitForExpect from "wait-for-expect";
+import { Provider } from "react-redux";
+import configureStore from "redux-mock-store";
 import UserSettingsPage from "./UserSettingsPage";
 
 jest.mock("axios");
+jest.mock("react-ga");
+
+const mockStore = configureStore();
 
 describe("UserSettingsPage", () => {
     afterEach(() => {
@@ -12,36 +16,43 @@ describe("UserSettingsPage", () => {
     });
 
     it("renders a LoadingSpinner until the query resolves", () => {
+        const store = mockStore({
+            userDetails: {
+                data: null,
+                isLoading: true,
+            },
+        });
+
         const wrapper = mount(
-            <UserSettingsPage isMobile={false} />,
+            <Provider store={store}>
+                <UserSettingsPage isMobile={false} />
+            </Provider>,
         );
 
         expect(wrapper.find("LoadingSpinner").length).toBe(1);
     });
 
     it("renders a GenericPage containing a UserSettings after receiving data", async () => {
-        axios.post.mockImplementationOnce(() => Promise.resolve({
-            data: {
+        const store = mockStore({
+            userDetails: {
                 data: {
-                    userDetails: {
-                        username: "username",
-                        sendReminderEmails: true,
-                    },
+                    username: "username",
+                    sendReminderEmails: true,
                 },
+                isLoading: false,
             },
-        }));
+        });
 
         const wrapper = mount(
-            <UserSettingsPage isMobile={false} />,
+            <Provider store={store}>
+                <UserSettingsPage isMobile={false} />
+            </Provider>,
         );
 
-        await waitForExpect(() => {
-            wrapper.update();
-            expect(wrapper.find("GenericPage").get(0).props.title).toBe("Settings");
-            expect(wrapper.find("GenericPage").get(0).props.isMobile).toBe(false);
-            expect(wrapper.find("GenericPage UserSettings").length).toBe(1);
-            expect(wrapper.find("GenericPage UserSettings").get(0).props.currentEmail).toBe("username");
-            expect(wrapper.find("GenericPage UserSettings").get(0).props.currentSendReminderEmails).toBe(true);
-        });
+        expect(wrapper.find("GenericPage").get(0).props.title).toBe("Settings");
+        expect(wrapper.find("GenericPage").get(0).props.isMobile).toBe(false);
+        expect(wrapper.find("GenericPage UserSettings").length).toBe(1);
+        expect(wrapper.find("GenericPage UserSettings").get(0).props.currentEmail).toBe("username");
+        expect(wrapper.find("GenericPage UserSettings").get(0).props.currentSendReminderEmails).toBe(true);
     });
 });
