+++ src/store.js
@@ -1,0 +1,10 @@
+import { configureStore } from "@reduxjs/toolkit";
+import userDetailsReducer from "./reducers/userDetailsSlice";
+
+const store = configureStore({
+    reducer: {
+        userDetails: userDetailsReducer,
+    },
+});
+
+export default store;
